+++ palu/model/svd_llama/__init__.py
@@ -1,6 +1,6 @@
 from transformers import AutoConfig, AutoModelForCausalLM, AutoTokenizer, LlamaTokenizer
-from .configuration_asvd_llama import PaluLlamaConfig
-from .modeling_asvd_llama import PaluLlamaForCausalLM
+from .configuration_palu_llama import PaluLlamaConfig
+from .modeling_palu_llama import PaluLlamaForCausalLM
 
 AutoConfig.register("palullama", PaluLlamaConfig)
 AutoModelForCausalLM.register(PaluLlamaConfig, PaluLlamaForCausalLM)
