+++ palu/model/svd_mistral/__init__.py
@@ -1,6 +1,6 @@
 from transformers import AutoConfig, AutoModelForCausalLM, AutoTokenizer, LlamaTokenizer
-from .configuration_asvd_mistral import PaluMistralConfig
-from .modeling_asvd_mistral import PaluMistralForCausalLM
+from .configuration_palu_mistral import PaluMistralConfig
+from .modeling_palu_mistral import PaluMistralForCausalLM
 
 AutoConfig.register("palumistral", PaluMistralConfig)
 AutoModelForCausalLM.register(PaluMistralConfig, PaluMistralForCausalLM)
